feat: rename and remove expressive from package name
