+++ CHANGELOG.md
@@ -22,13 +22,13 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#19](https://github.com/xtreamwayz/expressive-messenger/pull/19) fixes the version constrain for the event-dispatcher.
+- [#19](https://github.com/xtreamwayz/psr-container-messenger/pull/19) fixes the version constrain for the event-dispatcher.
 
 ## 1.0.0 - 2019-12-17
 
 ### Added
 
-- [#18](https://github.com/xtreamwayz/expressive-messenger/pull/18) adds support for messenger 4.4.
+- [#18](https://github.com/xtreamwayz/psr-container-messenger/pull/18) adds support for messenger 4.4.
 
 ### Changed
 
@@ -54,7 +54,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Changed
 
-- [#17](https://github.com/xtreamwayz/expressive-messenger/pull/17) removes default middleware config as it does not work as intended.
+- [#17](https://github.com/xtreamwayz/psr-container-messenger/pull/17) removes default middleware config as it does not work as intended.
 
 ### Deprecated
 
@@ -72,9 +72,9 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- [#14](https://github.com/xtreamwayz/expressive-messenger/pull/14) adds proper doctrine transport support.
+- [#14](https://github.com/xtreamwayz/psr-container-messenger/pull/14) adds proper doctrine transport support.
 
-- [#13](https://github.com/xtreamwayz/expressive-messenger/pull/13) adds optional logging to middleware and console commands.
+- [#13](https://github.com/xtreamwayz/psr-container-messenger/pull/13) adds optional logging to middleware and console commands.
 
 ### Changed
 
@@ -96,15 +96,15 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- [#10](https://github.com/xtreamwayz/expressive-messenger/pull/10) adds Symfony Messenger version 4.3.x support.
+- [#10](https://github.com/xtreamwayz/psr-container-messenger/pull/10) adds Symfony Messenger version 4.3.x support.
 
 ### Changed
 
-- [#10](https://github.com/xtreamwayz/expressive-messenger/pull/10) changed to the original consume messages command.
+- [#10](https://github.com/xtreamwayz/psr-container-messenger/pull/10) changed to the original consume messages command.
 
-- [#10](https://github.com/xtreamwayz/expressive-messenger/pull/10) changed the naming of supplied busses and related middleware.
+- [#10](https://github.com/xtreamwayz/psr-container-messenger/pull/10) changed the naming of supplied busses and related middleware.
 
-- [#10](https://github.com/xtreamwayz/expressive-messenger/pull/10) changed to the PHP internal serializer.
+- [#10](https://github.com/xtreamwayz/psr-container-messenger/pull/10) changed to the PHP internal serializer.
 
 ### Deprecated
 
@@ -112,13 +112,13 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Removed
 
-- [#10](https://github.com/xtreamwayz/expressive-messenger/pull/10) removed add default middleware config option. Middleware needs to be added manually now. The supplied busses already have the sender and handler middleware setup.
+- [#10](https://github.com/xtreamwayz/psr-container-messenger/pull/10) removed add default middleware config option. Middleware needs to be added manually now. The supplied busses already have the sender and handler middleware setup.
 
-- [#10](https://github.com/xtreamwayz/expressive-messenger/pull/10) removed obsolete symfony serializer factories.
+- [#10](https://github.com/xtreamwayz/psr-container-messenger/pull/10) removed obsolete symfony serializer factories.
 
-- [#10](https://github.com/xtreamwayz/expressive-messenger/pull/10) removed enqueue transport support as messenger supplies it own.
+- [#10](https://github.com/xtreamwayz/psr-container-messenger/pull/10) removed enqueue transport support as messenger supplies it own.
 
-- [#10](https://github.com/xtreamwayz/expressive-messenger/pull/10) removed deprecated LoggingMiddleware.
+- [#10](https://github.com/xtreamwayz/psr-container-messenger/pull/10) removed deprecated LoggingMiddleware.
 
 ### Fixed
 
@@ -144,7 +144,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#9](https://github.com/xtreamwayz/expressive-messenger/pull/9) restricts Symfony to version 4.2.x.
+- [#9](https://github.com/xtreamwayz/psr-container-messenger/pull/9) restricts Symfony to version 4.2.x.
 
 ## 0.4.1 - 2019-03-06
 
@@ -154,7 +154,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Changed
 
-- [#6](https://github.com/xtreamwayz/expressive-messenger/pull/6) uses container locators for handlers and senders.
+- [#6](https://github.com/xtreamwayz/psr-container-messenger/pull/6) uses container locators for handlers and senders.
 
 ### Deprecated
 
@@ -166,13 +166,13 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#6](https://github.com/xtreamwayz/expressive-messenger/pull/6) fixes senders being wrapped in callables where SenderInterface objects are expected.
+- [#6](https://github.com/xtreamwayz/psr-container-messenger/pull/6) fixes senders being wrapped in callables where SenderInterface objects are expected.
 
 ## 0.4.0 - 2018-12-31
 
 ### Added
 
-- [#5](https://github.com/xtreamwayz/expressive-messenger/pull/5) added messenger 4.2 compatibility.
+- [#5](https://github.com/xtreamwayz/psr-container-messenger/pull/5) added messenger 4.2 compatibility.
 
 ### Changed
 
@@ -221,21 +221,21 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#4](https://github.com/xtreamwayz/expressive-messenger/pull/4) adds the missing LoggingMiddlewareFactory.
+- [#4](https://github.com/xtreamwayz/psr-container-messenger/pull/4) adds the missing LoggingMiddlewareFactory.
 
 ## 0.3.0 - 2018-08-22
 
 ### Added
 
-- [#3](https://github.com/xtreamwayz/expressive-messenger/pull/3) adds message handling middleware which sends messages
+- [#3](https://github.com/xtreamwayz/psr-container-messenger/pull/3) adds message handling middleware which sends messages
   to multiple handlers.
 
-- [#3](https://github.com/xtreamwayz/expressive-messenger/pull/3) adds configuration for transports with DNS. This
+- [#3](https://github.com/xtreamwayz/psr-container-messenger/pull/3) adds configuration for transports with DNS. This
   enables support for all available enqueue transports.
 
 ### Changed
 
-- [#3](https://github.com/xtreamwayz/expressive-messenger/pull/3) changes how a message bus and enqueue transport is
+- [#3](https://github.com/xtreamwayz/psr-container-messenger/pull/3) changes how a message bus and enqueue transport is
   created. This comes with a major configuration overhaul.
 
 ### Deprecated
@@ -244,7 +244,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Removed
 
-- [#3](https://github.com/xtreamwayz/expressive-messenger/pull/3) removes the redis factory class. The preferred way is
+- [#3](https://github.com/xtreamwayz/psr-container-messenger/pull/3) removes the redis factory class. The preferred way is
   to use DNS to configure enqueue transports.
 
 ### Fixed
@@ -271,13 +271,13 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#2](https://github.com/xtreamwayz/expressive-messenger/pull/2) fixes the console command to be compatible with xtreamwayz/expressive-console.
+- [#2](https://github.com/xtreamwayz/psr-container-messenger/pull/2) fixes the console command to be compatible with xtreamwayz/expressive-console.
 
 ## 0.2.0 - 2018-08-02
 
 ### Added
 
-- [#1](https://github.com/xtreamwayz/expressive-messenger/pull/1) adds support for messenger 4.1.
+- [#1](https://github.com/xtreamwayz/psr-container-messenger/pull/1) adds support for messenger 4.1.
 
 ### Changed
