+++ LICENSE.md
@@ -1,4 +1,4 @@
-Copyright (c) 2018-2019 Geert Eltink
+Copyright (c) 2018-2020 Geert Eltink
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
