+++ README.md
@@ -1,13 +1,13 @@
-# Expressive Messenger
+# PSR Container Messenger
 
-_Message bus and queue for Zend Expressive with Symfony Messenger_
+_PSR-11 Container compatible Symfony Messenger message bus and queue._
 
-[![Docs Status](https://github.com/xtreamwayz/expressive-messenger/workflows/build-docs/badge.svg)](https://github.com/xtreamwayz/expressive-messenger/actions)
-[![Build Status](https://github.com/xtreamwayz/expressive-messenger/workflows/qa-tests/badge.svg)](https://github.com/xtreamwayz/expressive-messenger/actions)
-[![Downloads](https://img.shields.io/packagist/dt/xtreamwayz/expressive-messenger.svg)](https://packagist.org/packages/xtreamwayz/expressive-messenger)
-[![Packagist](https://img.shields.io/packagist/v/xtreamwayz/expressive-messenger.svg)](https://packagist.org/packages/xtreamwayz/expressive-messenger)
+[![Docs Status](https://github.com/xtreamwayz/psr-container-messenger/workflows/build-docs/badge.svg)](https://github.com/xtreamwayz/psr-container-messenger/actions)
+[![Build Status](https://github.com/xtreamwayz/psr-container-messenger/workflows/qa-tests/badge.svg)](https://github.com/xtreamwayz/psr-container-messenger/actions)
+[![Downloads](https://img.shields.io/packagist/dt/xtreamwayz/psr-container-messenger.svg)](https://packagist.org/packages/xtreamwayz/psr-container-messenger)
+[![Packagist](https://img.shields.io/packagist/v/xtreamwayz/psr-container-messenger.svg)](https://packagist.org/packages/xtreamwayz/psr-container-messenger)
 
-This packages brings message buses to your Zend Expressive project. Basically it's a bundle of factories to make
+This packages brings message buses to your PSR-11 container based project. It's a bundle of factories to make
 life easier for you. The real work is done by [Symfony Messenger](https://github.com/symfony/messenger).
 
 It comes with pre-configured command, event and query buses for your convenience. Or don't use them if you want to
@@ -16,14 +16,14 @@ create your own. Transports can be used to queue your messages or send and recei
 ## Installation
 
 ```bash
-composer require xtreamwayz/expressive-messenger
+composer require xtreamwayz/psr-container-messenger
 ```
 
 ## Documentation
 
 All project documentation is located in the [./docs](./docs) folder. If you would like to contribute
 to the documentation, please submit a pull request. You can read the docs online:
-https://xtreamwayz.netlify.com/expressive-messenger/
+https://xtreamwayz.netlify.com/psr-container-messenger/
 
 ## Contributing
