+++ example/single-bus-config.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace App;
 
-use Xtreamwayz\Expressive\Messenger\Container\HandleMessageMiddlewareFactory;
-use Xtreamwayz\Expressive\Messenger\Container\MessageBusFactory;
-use Xtreamwayz\Expressive\Messenger\Container\SendMessageMiddlewareFactory;
+use Xtreamwayz\PsrContainerMessenger\Container\HandleMessageMiddlewareFactory;
+use Xtreamwayz\PsrContainerMessenger\Container\MessageBusFactory;
+use Xtreamwayz\PsrContainerMessenger\Container\SendMessageMiddlewareFactory;
 
 return [
     // phpcs:disable
