+++ test/EventBusTest.php
@@ -2,15 +2,15 @@
 
 declare(strict_types=1);
 
-namespace XtreamwayzTest\Expressive\Messenger;
+namespace Xtreamwayz\PsrContainerMessenger\Test;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Messenger\MessageBus;
 use Symfony\Component\Messenger\MessageBusInterface;
-use Xtreamwayz\Expressive\Messenger\ConfigProvider;
-use XtreamwayzTest\Expressive\Messenger\Fixtures\DummyEvent;
-use XtreamwayzTest\Expressive\Messenger\Fixtures\DummyEventHandler;
-use XtreamwayzTest\Expressive\Messenger\Fixtures\DummyEventHandlerTwo;
+use Xtreamwayz\PsrContainerMessenger\ConfigProvider;
+use Xtreamwayz\PsrContainerMessenger\Test\Fixtures\DummyEvent;
+use Xtreamwayz\PsrContainerMessenger\Test\Fixtures\DummyEventHandler;
+use Xtreamwayz\PsrContainerMessenger\Test\Fixtures\DummyEventHandlerTwo;
 use Zend\ServiceManager\Config;
 use Zend\ServiceManager\ServiceManager;
 use function array_replace_recursive;
