+++ test/QueryBusTest.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace XtreamwayzTest\Expressive\Messenger;
+namespace Xtreamwayz\PsrContainerMessenger\Test;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Messenger\Exception\NoHandlerForMessageException;
 use Symfony\Component\Messenger\MessageBus;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\Stamp\HandledStamp;
-use Xtreamwayz\Expressive\Messenger\ConfigProvider;
-use XtreamwayzTest\Expressive\Messenger\Fixtures\DummyQuery;
-use XtreamwayzTest\Expressive\Messenger\Fixtures\DummyQueryHandler;
+use Xtreamwayz\PsrContainerMessenger\ConfigProvider;
+use Xtreamwayz\PsrContainerMessenger\Test\Fixtures\DummyQuery;
+use Xtreamwayz\PsrContainerMessenger\Test\Fixtures\DummyQueryHandler;
 use Zend\ServiceManager\Config;
 use Zend\ServiceManager\ServiceManager;
 use function array_replace_recursive;
